Rename: domain 하위 경로로 mindmap 도메인 경로 수정
