+++ backend/src/main/java/capstone/backend/domain/mindmap/dto/request/MindMapRequest.java
@@ -1,7 +1,7 @@
-package capstone.backend.mindmap.dto.request;
+package capstone.backend.domain.mindmap.dto.request;
 
-import capstone.backend.mindmap.entity.MindMapType;
-import capstone.backend.mindmap.entity.Node;
+import capstone.backend.domain.mindmap.entity.MindMapType;
+import capstone.backend.domain.mindmap.entity.Node;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
