+++ backend/src/main/java/capstone/backend/domain/mindmap/entity/MindMapType.java
@@ -1,4 +1,4 @@
-package capstone.backend.mindmap.entity;
+package capstone.backend.domain.mindmap.entity;
 
 public enum MindMapType {
     TODO,
