+++ backend/src/main/java/capstone/backend/global/redis/batch/MindMapPublisher.java
@@ -1,11 +1,7 @@
 package capstone.backend.global.redis.batch;
 
-import capstone.backend.mindmap.entity.MindMap;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
 @Service
