+++ backend/src/main/java/capstone/backend/global/security/SecurityConfig.java
@@ -34,7 +34,7 @@ public class SecurityConfig {
     private final JwtAuthenticationFilter jwtAuthenticationFilter;
 
     private static final String[] SWAGGER_ENDPOINTS = { "/swagger", "/swagger-ui/**", "/api-docs/**", "/v3/api-docs/**" };
-    private static final String[] PUBLIC_ENDPOINTS = { "/", "/api/auth/**" };
+    private static final String[] PUBLIC_ENDPOINTS = { "/", "/api/auth/**", "/config/**" };
     private static final String[] STATIC_ENDPOINTS = { "/error", "/favicon.ico", "/static/**",
             "/public/**", "/resources/**", "/META-INF/resources/**" };
